backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
fastapi==0.115.8
h11==0.14.0
idna==3.10
pydantic==2.10.6
pydantic_core==2.27.2
sniffio==1.3.1
starlette==0.45.3
typing_extensions==4.12.2
uvicorn==0.34.0
stripe==9.8.0
PyJWT==2.10.1
SQLAlchemy==1.4.32
python-dotenv==1.0.1
psycopg2-binary==2.9.10
